refactor: consolidate Pydantic domain models into src/domain/ (closes #52)
